requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.3
chromadb==0.5.23
sentence-transformers==3.3.1
pydantic==2.9.2
httpx==0.27.2



